refactor(cli): move whole cli over to telemetry api
